Rename google_oauth2 -> default_auth_provider
